+++ components/vf-banner/CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.5.0
+
+- centralises logic to "close" a banner
+- adds padding to the document to accomodate a fixed-position banner so it won't hide content
+  - https://github.com/visual-framework/vf-core/issues/1119
+
 ### 1.4.0
 
 - removes inline padding on `vf-banner` class